requirements.txt
pyyaml
requests
urllib3
nose
coverage